+++ upload/library/Iversia/FAQ/FileSums.php
@@ -1,0 +1,39 @@
+<?php
+
+class Iversia_FAQ_FileSums
+{
+	public static function getHashes()
+	{
+		return array (
+  'library/Iversia/FAQ/AlertHandler/Question.php' => '1e9fc9156b156d617c2a35251030b35d',
+  'library/Iversia/FAQ/AttachmentHandler/Question.php' => '88e62068423f5cdaea86760d1e747320',
+  'library/Iversia/FAQ/BbCode/Formatter/Base.php' => '6d913dbe2f817693c58078082b06c4b8',
+  'library/Iversia/FAQ/ControllerHelper/FAQ.php' => '1b20912c169da2ddf5bcda24e525e63e',
+  'library/Iversia/FAQ/ControllerPublic/FAQ.php' => '31c045feecc988dcbdf7f053637574f3',
+  'library/Iversia/FAQ/CronEntry/FAQ.php' => '751e2e738067a54ccbe9ffb4cc8490c0',
+  'library/Iversia/FAQ/DataWriter/Category.php' => 'dff1df1343252504723190a7fc4184dc',
+  'library/Iversia/FAQ/DataWriter/Question.php' => '50c741adcddced8e05915dd226b3ddd7',
+  'library/Iversia/FAQ/Installer.php' => '00311bf7a4cceb86189b77b821a1e239',
+  'library/Iversia/FAQ/LikeHandler/Question.php' => 'd4cc23393226963185f988f0069187ae',
+  'library/Iversia/FAQ/Listener/BbCode.php' => '0aa67f16c383d8bdf338ad13b95030ef',
+  'library/Iversia/FAQ/Listener/Extend.php' => 'f629a48d9f25ba9eb166ebc74c1746ee',
+  'library/Iversia/FAQ/Listener/TemplateHook.php' => 'fef7ad786e6f6b3dae54ddb8553cded7',
+  'library/Iversia/FAQ/Listener/Widgets.php' => '1e71b82710b21a807b8bb00b80aeb90b',
+  'library/Iversia/FAQ/Model/Category.php' => '48058aba484392777ab620e3203dd1ec',
+  'library/Iversia/FAQ/Model/Question.php' => 'e85cb4ce2f076950a4a2924b95aa7c67',
+  'library/Iversia/FAQ/Model/Search.php' => '101773cc037f48846d3c3ad93b65fcc8',
+  'library/Iversia/FAQ/NewsFeedHandler/Question.php' => 'c17cc8f1e42df8f956896adcb5aad652',
+  'library/Iversia/FAQ/Route/Prefix/FAQ.php' => '0a33b86424fc9efe06bafa39074eaff5',
+  'library/Iversia/FAQ/Search/DataHandler/Question.php' => '556f0d68e6c5510aafbe03754fffb2e5',
+  'library/Iversia/FAQ/ViewPublic/Create.php' => '45c98f4e2cf83cbca7028cb3bc1245a4',
+  'library/Iversia/FAQ/ViewPublic/Edit.php' => '4bc6ec02d010def842778a15a6d47e60',
+  'library/Iversia/FAQ/ViewPublic/Index.php' => 'd4d27633d7ab9e90e8cc4d4eff78a630',
+  'library/Iversia/FAQ/ViewPublic/LatestAnswers.php' => 'a9e90aee6acade6f03d27e590bb539ca',
+  'library/Iversia/FAQ/ViewPublic/LikeConfirmed.php' => 'f651ae599ae8b28a8c0be27a4696dc6c',
+  'library/Iversia/FAQ/ViewPublic/Permalink.php' => 'd169106bd4b14691f94f32ab33dedb88',
+  'library/Iversia/FAQ/WidgetRenderer/LatestAnswers.php' => '7329feb5e7e5446846a8d34302fc779b',
+  'library/Iversia/FAQ/WidgetRenderer/MostPopular.php' => '9874beea8a6be44222b1494a827061eb',
+  'library/Iversia/FAQ/WidgetRenderer/StickyAnswers.php' => '96a0bcc7220ce34a4092b5b07774d971',
+);
+	}
+}